+++ src/components/ProductListing/filter-components/index.js
@@ -1,0 +1,4 @@
+export { FilterHeader } from "./filterHeader";
+export { SortByFilter } from "./sortBy";
+export { CategoryFilter } from "./categoryFilter";
+export { BrandFilter } from "./brandFilter";
